abstract extractPreview code and reuse for printview
